+++ lib/dart/MemoryState.dart
@@ -37,6 +37,8 @@ class MemoryState {
   String username = "";
   String password = "";
   String applicationVersion = "";
+  bool useCustomFont = false;
+  String customFontFamily = "";
 
   List<int> AlterBackgroundColor = [255, 90, 90, 90];
   List<int> AlterTextColor = [179, 255, 255, 255];