+++ lib/dart/globalHelpers.dart
@@ -1,7 +1,21 @@
+import 'dart:math';
+
 bool identicalColors(List<int> a, List<int> b) {
   if (a[0] != b[0] || a[1] != b[1] || a[2] != b[2] || a[3] != b[3]) {
     return false;
   } else {
     return true;
   }
+}
+
+// Define a reusable function
+String generateRandomString(int length) {
+  final random = Random();
+  const availableChars = 'AaBbCcDdEeFfGgHhIiJjKkLlMmNnOoPpQqRrSsTtUuVvWwXxYyZz';
+  final randomString = List.generate(
+    length,
+    (index) => availableChars[random.nextInt(availableChars.length)],
+  ).join();
+
+  return randomString;
 }