+++ lib/sb.dart
@@ -1,8 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:switchboard/dart/MemoryState.dart';
+import 'package:switchboard/globalHelpers.dart';
 import 'package:switchboard/pages/aboutPage.dart';
 import 'package:switchboard/pages/accountPage.dart';
 import 'package:switchboard/pages/editAlter.dart';
 import 'package:switchboard/pages/editFields.dart';
+import 'package:switchboard/pages/fontPage.dart';
 import 'package:switchboard/pages/loginPage.dart';
 import 'package:switchboard/pages/privacyPolicyPage.dart';
 import 'package:switchboard/pages/registerPage.dart';
@@ -13,19 +16,25 @@ class Switchboard extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: "Switchboard",
-      theme: ThemeData.dark(),
-      routes: {
-        "/": (ctx) => SBLoginPage(),
-        "/about": (ctx) => SBAboutPage(),
-        "/register": (ctx) => SBRegisterPage(),
-        "/privacy": (ctx) => PrivacyPolicyPage(),
-        "/account": (ctx) => AccountPage(),
-        "/account/settings": (ctx) => SettingsPage(),
-        "/account/settings/fields": (ctx) => EditFieldsPage(),
-        "/account/settings/fields/edit": (ctx) => EditField(),
-        "/editAlter": (ctx) => EditAlterPage(),
+    return ValueListenableBuilder(
+      valueListenable: customFontNotifier,
+      builder: (ctx, value, _) {
+        return MaterialApp(
+          title: "Switchboard",
+          theme: getApplicationTheme(),
+          routes: {
+            "/": (ctx) => SBLoginPage(),
+            "/about": (ctx) => SBAboutPage(),
+            "/register": (ctx) => SBRegisterPage(),
+            "/privacy": (ctx) => PrivacyPolicyPage(),
+            "/account": (ctx) => AccountPage(),
+            "/account/settings": (ctx) => SettingsPage(),
+            "/account/settings/font": (ctx) => FontPage(),
+            "/account/settings/fields": (ctx) => EditFieldsPage(),
+            "/account/settings/fields/edit": (ctx) => EditField(),
+            "/editAlter": (ctx) => EditAlterPage(),
+          },
+        );
       },
     );
   }